pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pylect"
version = "0.3.0"
dependencies = [
    "black",
    "isort",
    "pylint",
    "pyperclip",
    "pytest",
    "python-dateutil",
    "python-dotenv",
    "requests",
]
authors = [{ name = "William Wingate", email = "wtwingate@gmail.com" }]
description = "Tools for working with lectionaries and liturgical calendars."
readme = "README.md"
license = { file = "LICENSE.txt" }

[project.urls]
Repository = "https://github.com/wtwingate/pylect"

[project.scripts]
pylect = "pylect.cli:start"

[tool.black]
line-length = 79
